Create a mockable DI container to clean up test code
